add sstable builder
